ds_studio_llama/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
python-dotenv==1.0.1
boto3==1.35.81
botocore==1.35.81
certifi==2024.12.14
cffi==1.17.1

langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.4
langdetect==1.0.9
langsmith==0.2.10
lazy_loader==0.4
llama-cloud==0.1.8
llama-index==0.11.2
llama-index-agent-openai==0.3.3
llama-index-cli==0.3.1
llama-index-core==0.11.2
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.3.1
llama-index-legacy==0.9.48.post4
llama-index-llms-anthropic==0.2.1
llama-index-llms-bedrock==0.2.6
llama-index-llms-bedrock-converse==0.3.9
llama-index-llms-openai==0.2.2
llama-index-multi-modal-llms-openai==0.2.1
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.19
pandas==2.2.0
openpyxl==3.1.5
python-docx==1.1.2
PyPDF2==3.0.1
pypdf==4.0.2
sentence-transformers>=2.6.0
torch>=2.2.0
langchain>=0.3.14
langchain-community>=0.3.14
transformers>=4.38.2
pdf2image==1.16.3
Pillow==9.5.0
docx2pdf==0.1.8
docx2txt==0.8
python-pptx==0.6.22
unstructured>=0.11.0
unstructured[pptx]>=0.11.0
pdfminer.six>=20221105
